Refactor hook
