+++ src/components/spatial/SpatialSubscriber.js
@@ -1,7 +1,9 @@
 import React, { useEffect, useMemo, useCallback } from 'react';
 import TitleInfo from '../TitleInfo';
 import { capitalize } from '../../utils';
-import { useDeckCanvasSize, useReady, useUrls } from '../hooks';
+import {
+  useDeckCanvasSize, useReady, useUrls, useExpressionValueGetter,
+} from '../hooks';
 import { setCellSelection, mergeCellSets } from '../utils';
 import {
   useCellsData,
@@ -280,29 +282,9 @@ export default function SpatialSubscriber(props) {
     locationsCount,
   });
 
-  // Get a mapping from cell ID to row index in the gene expression matrix.
-  const cellIdMap = useMemo(() => {
-    const result = {};
-    if (attrs && attrs.rows) {
-      // eslint-disable-next-line no-plusplus
-      for (let i = 0; i < attrs.rows.length; i++) {
-        result[attrs.rows[i]] = i;
-      }
-    }
-    return result;
-  }, [attrs]);
-
   // Set up a getter function for gene expression values, to be used
   // by the DeckGL layer to obtain values for instanced attributes.
-  const getExpressionValue = useCallback((entry) => {
-    const cellId = entry[0];
-    if (cellIdMap && expressionData && expressionData[0]) {
-      const cellIndex = cellIdMap[cellId];
-      const val = expressionData[0][cellIndex];
-      return val;
-    }
-    return 0;
-  }, [cellIdMap, expressionData]);
+  const getExpressionValue = useExpressionValueGetter({ attrs, expressionData });
 
   return (
     <TitleInfo
